01.07.2021

+++ src/main/java/de/netempire/movie_library/jpa/MovieRepository.java
@@ -7,5 +7,5 @@ public interface MovieRepository extends JpaRepository<Movie, Long> {
 
     Movie findByTitle(String title);
 
-    Movie findAllById(int i);
+    Movie findAllById(long i);
 }
